+++ example/src/screens/PaymentSheetWithPmoSfuScreen.tsx
@@ -94,7 +94,7 @@ export default function PaymentSheetWithPmoSfuScreen() {
 
       const address: Address = {
         city: 'San Francisco',
-        country: 'AT',
+        country: 'US',
         line1: '510 Townsend St.',
         line2: '123 Street',
         postalCode: '94102',