+++ ios/NewArch/ApplePayButtonComponentView.mm
@@ -98,6 +98,12 @@ using namespace facebook::react;
   _view.disabled = newViewProps.disabled;
   _view.borderRadius = @(newViewProps.borderRadius);
 
+  // Set the boolean flags from props
+  _view.hasShippingMethodCallback = newViewProps.hasShippingMethodCallback;
+  _view.hasShippingContactCallback = newViewProps.hasShippingContactCallback;
+  _view.hasCouponCodeCallback = newViewProps.hasCouponCodeCallback;
+  _view.hasOrderTrackingCallback = newViewProps.hasOrderTrackingCallback;
+
   [super updateProps:props oldProps:oldProps];
 
   [_view didSetProps];