+++ src/functions.ts
@@ -185,7 +185,10 @@ export const handleNextAction = async (
             paymentIntentClientSecret,
             returnURL ?? null
           )
-        : await NativeStripeSdk.handleNextAction(paymentIntentClientSecret);
+        : await NativeStripeSdk.handleNextAction(
+            paymentIntentClientSecret,
+            returnURL ?? null
+          );
     if (error) {
       return {
         error,